Fix StyleCop SA1402: A C# document may only contain a single class at the root level unless all of the classes are partial and are of the same type.
