+++ src/System.Web.OData/OData/Query/Expressions/SelectExpandWrapper.cs
@@ -8,7 +8,6 @@ using System.Web.OData.Formatter;
 using System.Web.OData.Formatter.Serialization;
 using System.Web.OData.Properties;
 using Microsoft.OData.Edm;
-using Newtonsoft.Json;
 
 namespace System.Web.OData.Query.Expressions
 {
@@ -130,28 +129,6 @@ namespace System.Web.OData.Query.Expressions
             Contract.Assert(ModelID != null);
 
             return ModelContainer.GetModel(ModelID);
-        }
-    }
-
-    /// <summary>
-    /// Represents a container class that contains properties that are either selected or expanded using $select and $expand.
-    /// </summary>
-    /// <typeparam name="TElement">The element being selected and expanded.</typeparam>
-    [JsonConverter(typeof(SelectExpandWrapperConverter))]
-    internal class SelectExpandWrapper<TElement> : SelectExpandWrapper
-    {
-        /// <summary>
-        /// Gets or sets the instance of the element being selected and expanded.
-        /// </summary>
-        public TElement Instance
-        {
-            get { return (TElement)UntypedInstance; }
-            set { UntypedInstance = value; }
-        }
-
-        protected override Type GetElementType()
-        {
-            return UntypedInstance == null ? typeof(TElement) : UntypedInstance.GetType();
         }
     }
 }
